+++ app/src/main/java/org/linphone/activities/main/chat/adapters/ChatMessagesListAdapter.kt
@@ -452,7 +452,11 @@ private class ChatMessageDiffCallback : DiffUtil.ItemCallback<EventLogData>() {
         newItem: EventLogData
     ): Boolean {
         return if (newItem.eventLog.type == EventLog.Type.ConferenceChatMessage) {
-            newItem.eventLog.chatMessage?.state == ChatMessage.State.Displayed
+            val oldData = (oldItem.data as ChatMessageData)
+            val newData = (newItem.data as ChatMessageData)
+            val previous = oldData.hasPreviousMessage == newData.hasPreviousMessage
+            val next = oldData.hasNextMessage == newData.hasNextMessage
+            newItem.eventLog.chatMessage?.state == ChatMessage.State.Displayed && previous && next
         } else true
     }
 }