+++ app/src/main/java/org/linphone/activities/main/chat/data/ChatMessageData.kt
@@ -59,6 +59,9 @@ class ChatMessageData(val chatMessage: ChatMessage) : GenericContactData(chatMes
 
     val replyData = MutableLiveData<ChatMessageData>()
 
+    var hasPreviousMessage = false
+    var hasNextMessage = false
+
     private var countDownTimer: CountDownTimer? = null
 
     private val listener = object : ChatMessageListenerStub() {
@@ -106,6 +109,11 @@ class ChatMessageData(val chatMessage: ChatMessage) : GenericContactData(chatMes
     }
 
     fun updateBubbleBackground(hasPrevious: Boolean, hasNext: Boolean) {
+        hasPreviousMessage = hasPrevious
+        hasNextMessage = hasNext
+        hideTime.value = false
+        hideAvatar.value = false
+
         if (hasPrevious) {
             hideTime.value = true
         }