+++ app/src/main/java/org/linphone/activities/main/chat/viewmodels/ChatRoomViewModel.kt
@@ -113,7 +113,7 @@ class ChatRoomViewModel(val chatRoom: ChatRoom) : ViewModel(), ContactDataInterf
         override fun onContactsUpdated() {
             Log.i("[Chat Room] Contacts have changed")
             contactLookup()
-            lastMessageText.value = formatLastMessage(chatRoom.lastMessageInHistory)
+            updateLastMessageToDisplay()
         }
     }
 
@@ -199,7 +199,7 @@ class ChatRoomViewModel(val chatRoom: ChatRoom) : ViewModel(), ContactDataInterf
 
         override fun onEphemeralMessageDeleted(chatRoom: ChatRoom, eventLog: EventLog) {
             Log.i("[Chat Room] Ephemeral message deleted, updated last message displayed")
-            lastMessageText.value = formatLastMessage(chatRoom.lastMessageInHistory)
+            updateLastMessageToDisplay()
         }
 
         override fun onEphemeralEvent(chatRoom: ChatRoom, eventLog: EventLog) {
@@ -226,7 +226,7 @@ class ChatRoomViewModel(val chatRoom: ChatRoom) : ViewModel(), ContactDataInterf
 
         contactLookup()
         updateParticipants()
-        lastMessageText.value = formatLastMessage(chatRoom.lastMessageInHistory)
+        updateLastMessageToDisplay()
 
         callInProgress.value = chatRoom.core.callsNb > 0
         updateRemotesComposing()
@@ -274,6 +274,10 @@ class ChatRoomViewModel(val chatRoom: ChatRoom) : ViewModel(), ContactDataInterf
         if (address != null) {
             coreContext.startCall(address)
         }
+    }
+
+    private fun updateLastMessageToDisplay() {
+        lastMessageText.value = formatLastMessage(chatRoom.lastMessageInHistory)
     }
 
     private fun formatLastMessage(msg: ChatMessage?): String {